+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Verb Changelog
 ## master
-- Do not automatically add `Accept-Charset`, `Accept-Encoding` and `Extension` headers to requests anymore.
+- Do not automatically add `Accept-Charset` header to requests anymore.
+- Fix `Accept` header being sent duplicated if included in a request specification.
 
 ## **2.2.0** - 2020-01-24
 - Renamed `verb-mode-prefix-map` to `verb-command-map` (existing `init.el` configs must be updated).