+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix `Accept` header being sent duplicated if included in a request specification.
 - Log a warning to `*Verb Log*` buffer when headers might get duplicated by url.el.
 - Overriding HTTP headers now ignores case (e.g. it is now possible to override `content-type` using `Content-Type`).
+- Added customizable option `verb-base-headers`.
 
 ## **2.2.0** - 2020-01-24
 - Renamed `verb-mode-prefix-map` to `verb-command-map` (existing `init.el` configs must be updated).